+++ src/MainSequence.cpp
@@ -1012,6 +1012,7 @@ void MainSequence::UpdateAgeAfterMassLoss() {
     double tMSprime  = MainSequence::CalculateLifetimeOnPhase(m_Mass, tBGBprime);
     
     m_Age *= tMSprime / tMS;
+    CalculateTimescales(m_Mass, m_Timescales);                                      // must update timescales
 }
 
 
@@ -1196,4 +1197,167 @@ void MainSequence::UpdateAfterMerger(double p_Mass, double p_HydrogenMass) {
     UpdateAttributesAndAgeOneTimestep(0.0, 0.0, 0.0, true);
     
     #undef timescales
+}
+
+
+
+/* 
+ * Interpolate Ge+ Critical Mass Ratios, for H-rich stars
+ * 
+ * Function takes input QCRIT_PRESCRIPTION, currently either of the prescriptions for critical mass ratios
+ * from Ge et al. (2020), GE or GE_IC. The first is the full adiabatic response, the second assumes
+ * artificially isentropic envelopes. From private communication with Ge, we have an updated datatable that
+ * includes qCrit for fully conservative and fully non-conservative MT, so we now interpolate on those as well.
+ *
+ * Interpolation is done linearly in logM, logR, and logZ
+ * 
+ * double BaseStar::InterpolateGeEtAlQCrit(const QCRIT_PRESCRIPTION p_qCritPrescription, const double p_massTransferEfficiencyBeta) 
+ * 
+ * @param   [IN]    p_qCritPrescription          Adopted critical mass ratio prescription
+ * @param   [IN]    p_massTransferEfficiencyBeta Mass transfer accretion efficiency
+ * @return                                       Interpolated value of either the critical mass ratio or zeta for given stellar mass / radius
+ */ 
+double MainSequence::InterpolateGeEtAlQCrit(const QCRIT_PRESCRIPTION p_qCritPrescription, const double p_massTransferEfficiencyBeta) {
+
+    // Iterate over the two QCRIT_GE tables to get the qcrits at each metallicity
+    double qCritPerMetallicity[2];
+    std::vector<GE_QCRIT_TABLE> qCritTables = { QCRIT_GE_LOW_Z, QCRIT_GE_HIGH_Z };
+
+    for (int ii=0; ii<2; ii++) { // iterate over the vector of tables to store qCrits per metallicity
+
+        // Get vector of masses from qCritTable
+        GE_QCRIT_TABLE &qCritTable = qCritTables[ii];
+        DBL_VECTOR massesFromQCritTable = std::get<0>(qCritTable);
+        GE_QCRIT_RADII_QCRIT_VECTOR radiiQCritsFromQCritTable = std::get<1>(qCritTable);
+
+        INT_VECTOR indices = utils::BinarySearch(massesFromQCritTable, m_Mass);
+        int lowerMassIndex = indices[0];
+        int upperMassIndex = indices[1];
+    
+        if (lowerMassIndex == -1) {                                                   // if masses are out of range, set to endpoints
+            lowerMassIndex = 0; 
+            upperMassIndex = 1;
+        } 
+        else if (upperMassIndex == -1) { 
+            lowerMassIndex = massesFromQCritTable.size() - 2; 
+            upperMassIndex = massesFromQCritTable.size() - 1;
+        } 
+    
+        // Get vector of radii from qCritTable for the lower and upper mass indices
+        std::vector<double> logRadiusVectorLowerMass = std::get<0>(radiiQCritsFromQCritTable[lowerMassIndex]);
+        std::vector<double> logRadiusVectorUpperMass = std::get<0>(radiiQCritsFromQCritTable[upperMassIndex]);
+    
+        // Get the qCrit vector for the lower and upper mass bounds 
+        std::vector<double> qCritVectorUpperEffLowerMass;
+        std::vector<double> qCritVectorUpperEffUpperMass;
+        std::vector<double> qCritVectorLowerEffLowerMass;
+        std::vector<double> qCritVectorLowerEffUpperMass;
+        
+        // Set the appropriate qCrit vector, depends on MT eff and whether you use GE STD or IC
+        if (p_qCritPrescription == QCRIT_PRESCRIPTION::GE) {
+            if (p_massTransferEfficiencyBeta > 0.5) {
+                qCritVectorUpperEffLowerMass = std::get<1>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorUpperEffUpperMass = std::get<1>(radiiQCritsFromQCritTable[upperMassIndex]);
+                qCritVectorLowerEffLowerMass = std::get<2>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorLowerEffUpperMass = std::get<2>(radiiQCritsFromQCritTable[upperMassIndex]);
+            }
+            else {
+                qCritVectorUpperEffLowerMass = std::get<2>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorUpperEffUpperMass = std::get<2>(radiiQCritsFromQCritTable[upperMassIndex]);
+                qCritVectorLowerEffLowerMass = std::get<3>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorLowerEffUpperMass = std::get<3>(radiiQCritsFromQCritTable[upperMassIndex]);
+
+            }
+        }
+        else if (p_qCritPrescription == QCRIT_PRESCRIPTION::GE_IC) {
+            if (p_massTransferEfficiencyBeta > 0.5) {
+                qCritVectorUpperEffLowerMass = std::get<4>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorUpperEffUpperMass = std::get<4>(radiiQCritsFromQCritTable[upperMassIndex]);
+                qCritVectorLowerEffLowerMass = std::get<5>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorLowerEffUpperMass = std::get<5>(radiiQCritsFromQCritTable[upperMassIndex]);
+            }
+            else {
+                qCritVectorUpperEffLowerMass = std::get<5>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorUpperEffUpperMass = std::get<5>(radiiQCritsFromQCritTable[upperMassIndex]);
+                qCritVectorLowerEffLowerMass = std::get<6>(radiiQCritsFromQCritTable[lowerMassIndex]);
+                qCritVectorLowerEffUpperMass = std::get<6>(radiiQCritsFromQCritTable[upperMassIndex]);
+
+            }
+        }
+    
+        // Get vector of radii from qCritTable for both lower and upper masses
+        INT_VECTOR indicesR0          = utils::BinarySearch(logRadiusVectorLowerMass, log10(m_Radius));
+        int lowerRadiusLowerMassIndex = indicesR0[0];
+        int upperRadiusLowerMassIndex = indicesR0[1];
+    
+        if (lowerRadiusLowerMassIndex == -1) {                                        // if radii are out of range, set to endpoints
+            lowerRadiusLowerMassIndex = 0; 
+            upperRadiusLowerMassIndex = 1; 
+        }
+        else if (upperRadiusLowerMassIndex == -1) {                                                   
+            lowerRadiusLowerMassIndex = logRadiusVectorLowerMass.size() - 2; 
+            upperRadiusLowerMassIndex = logRadiusVectorLowerMass.size() - 1; 
+        }
+    
+        INT_VECTOR indicesR1          = utils::BinarySearch(logRadiusVectorUpperMass, log10(m_Radius));
+        int lowerRadiusUpperMassIndex = indicesR1[0];
+        int upperRadiusUpperMassIndex = indicesR1[1];
+    
+        if (lowerRadiusUpperMassIndex == -1) {                                        // if radii are out of range, set to endpoints
+            lowerRadiusUpperMassIndex = 0; 
+            upperRadiusUpperMassIndex = 1; 
+        }
+        else if (upperRadiusUpperMassIndex == -1) {                                                   
+            lowerRadiusUpperMassIndex = logRadiusVectorUpperMass.size() - 2; 
+            upperRadiusUpperMassIndex = logRadiusVectorUpperMass.size() - 1; 
+        }
+    
+        // Set the 4 boundary points for the 2D interpolation
+        double qUppLowLow = qCritVectorUpperEffLowerMass[lowerRadiusLowerMassIndex];
+        double qUppLowUpp = qCritVectorUpperEffLowerMass[upperRadiusLowerMassIndex];
+        double qUppUppLow = qCritVectorUpperEffUpperMass[lowerRadiusUpperMassIndex];
+        double qUppUppUpp = qCritVectorUpperEffUpperMass[upperRadiusUpperMassIndex];
+        double qLowLowLow = qCritVectorLowerEffLowerMass[lowerRadiusLowerMassIndex];
+        double qLowLowUpp = qCritVectorLowerEffLowerMass[upperRadiusLowerMassIndex];
+        double qLowUppLow = qCritVectorLowerEffUpperMass[lowerRadiusUpperMassIndex];
+        double qLowUppUpp = qCritVectorLowerEffUpperMass[upperRadiusUpperMassIndex];
+    
+        double lowerLogRadiusLowerMass = logRadiusVectorLowerMass[lowerRadiusLowerMassIndex];
+        double upperLogRadiusLowerMass = logRadiusVectorLowerMass[upperRadiusLowerMassIndex];
+        double lowerLogRadiusUpperMass = logRadiusVectorUpperMass[lowerRadiusUpperMassIndex];
+        double upperLogRadiusUpperMass = logRadiusVectorUpperMass[upperRadiusUpperMassIndex];
+
+        double logLowerMass   = log10(massesFromQCritTable[lowerMassIndex]);
+        double logUpperMass   = log10(massesFromQCritTable[upperMassIndex]);
+    
+        // Interpolate on logR first, then logM, then on the efficiency, using nearest neighbor for extrapolation
+        double logRadius = log10(m_Radius);
+        double qCritUpperEffLowerMass = (logRadius < lowerLogRadiusLowerMass) ? qUppLowLow
+                                      : (logRadius > upperLogRadiusLowerMass) ? qUppLowUpp
+                                      : qUppLowLow + (upperLogRadiusLowerMass - logRadius) / (upperLogRadiusLowerMass - lowerLogRadiusLowerMass) * (qUppLowUpp - qUppLowLow);
+        double qCritUpperEffUpperMass = (logRadius < lowerLogRadiusUpperMass) ? qUppUppLow
+                                      : (logRadius > upperLogRadiusUpperMass) ? qUppUppUpp
+                                      : qUppUppLow + (upperLogRadiusUpperMass - logRadius) / (upperLogRadiusUpperMass - lowerLogRadiusUpperMass) * (qUppUppUpp - qUppUppLow);
+        double qCritLowerEffLowerMass = (logRadius < lowerLogRadiusLowerMass) ? qLowLowLow
+                                      : (logRadius > upperLogRadiusLowerMass) ? qLowLowUpp
+                                      : qLowLowLow + (upperLogRadiusLowerMass - logRadius) / (upperLogRadiusLowerMass - lowerLogRadiusLowerMass) * (qLowLowUpp - qLowLowLow);
+        double qCritLowerEffUpperMass = (logRadius < lowerLogRadiusUpperMass) ? qLowUppLow
+                                      : (logRadius > upperLogRadiusUpperMass) ? qLowUppUpp
+                                      : qLowUppLow + (upperLogRadiusUpperMass - logRadius) / (upperLogRadiusUpperMass - lowerLogRadiusUpperMass) * (qLowUppUpp - qLowUppLow);
+    
+        double logMass = log10(m_Mass);
+        double interpolatedQCritUpperEff = (logMass < logLowerMass) ? qCritUpperEffLowerMass
+                                         : (logMass > logUpperMass) ? qCritUpperEffUpperMass
+                                         : qCritUpperEffLowerMass + (logUpperMass - logMass) / (logUpperMass - logLowerMass) * (qCritUpperEffUpperMass - qCritUpperEffLowerMass);
+        double interpolatedQCritLowerEff = (logMass < logLowerMass) ? qCritLowerEffLowerMass
+                                         : (logMass > logUpperMass) ? qCritLowerEffUpperMass
+                                         : qCritLowerEffLowerMass + (logUpperMass - logMass) / (logUpperMass - logLowerMass) * (qCritLowerEffUpperMass - qCritLowerEffLowerMass);
+    
+        double interpolatedQCritForZ = p_massTransferEfficiencyBeta * interpolatedQCritUpperEff + (1.0 - p_massTransferEfficiencyBeta) * interpolatedQCritLowerEff;                 // Don't need to use nearest neighbor for this, beta is always between 0 and 1
+        qCritPerMetallicity[ii] = interpolatedQCritForZ;
+    }
+    double logZlo = -3;         // log10(0.001)
+    double logZhi = LOG10_ZSOL; // log10(0.02) 
+    
+    return qCritPerMetallicity[1] + (m_Log10Metallicity - logZhi)*(qCritPerMetallicity[1] - qCritPerMetallicity[0])/(logZhi - logZlo);
 }